refactor: create folder to impove file structure
