+++ src/commands/basicOperations/cat.js
